Moving to its respective folders
