+++ app/screens/integration_selector/selected_option/index.tsx
